Update caldgemm_cuda.cu
Removed tranpose because it creates a performance problem with overlapping copy and DGEMM kernels (the transpose kernel becomes blocked until previous DGEMM is almost finished, and then the Copy C HostToDevice is delayed so it cannot overlap DGEMM).  

Other possible solutions that could be investigated in the future:  
1.) change the order of operations, Copy C H2D first before calling transposes, (for example: copy A, copy B, copy C, transpose A, transpose B, DGEMM, copy C) then the copy will overlap with previous DGEMM.
2.) Use a "High Priority" cuda stream for transpose, so it will not wait for previous DGEMM thread blocks to be scheduled first.  

In my test this gives ~15% increased HPL performance.